[NU-1443] Tapir services: packages refactor
