+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/NodesApiHttpService.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.ui.services
+package pl.touk.nussknacker.ui.api
 
 import cats.data.EitherT
 import com.typesafe.scalalogging.LazyLogging
@@ -12,27 +12,10 @@ import pl.touk.nussknacker.restmodel.definition.UIValueParameter
 import pl.touk.nussknacker.restmodel.scenariodetails.ScenarioWithDetails
 import pl.touk.nussknacker.ui.UnauthorizedError
 import pl.touk.nussknacker.ui.additionalInfo.AdditionalInfoProviders
-import pl.touk.nussknacker.ui.api.NodesApiEndpoints
-import pl.touk.nussknacker.ui.api.NodesApiEndpoints.Dtos
-import pl.touk.nussknacker.ui.api.NodesApiEndpoints.Dtos.NodesError.{
-  MalformedTypingResult,
-  NoPermission,
-  NoProcessingType,
-  NoScenario
-}
-import pl.touk.nussknacker.ui.api.NodesApiEndpoints.Dtos.{
-  ExpressionSuggestionDto,
-  NodeValidationRequest,
-  NodeValidationRequestDto,
-  NodeValidationResult,
-  NodeValidationResultDto,
-  NodesError,
-  ParametersValidationRequest,
-  ParametersValidationRequestDto,
-  ParametersValidationResultDto,
-  decodeVariableTypes,
-  prepareTypingResultDecoder
-}
+import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints
+import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints.Dtos
+import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints.Dtos.NodesError.{MalformedTypingResult, NoPermission, NoProcessingType, NoScenario}
+import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints.Dtos.{ExpressionSuggestionDto, NodeValidationRequest, NodeValidationRequestDto, NodeValidationResult, NodeValidationResultDto, NodesError, ParametersValidationRequest, ParametersValidationRequestDto, ParametersValidationResultDto, decodeVariableTypes, prepareTypingResultDecoder}
 import pl.touk.nussknacker.ui.process.ProcessService
 import pl.touk.nussknacker.ui.process.ProcessService.GetScenarioWithDetailsOptions
 import pl.touk.nussknacker.ui.process.processingtype.ProcessingTypeDataProvider
