+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ScenarioParametersApiHttpService.scala
@@ -1,8 +1,8 @@
-package pl.touk.nussknacker.ui.services
+package pl.touk.nussknacker.ui.api
 
 import com.typesafe.scalalogging.LazyLogging
-import pl.touk.nussknacker.ui.api.ScenarioParametersApiEndpoints
-import pl.touk.nussknacker.ui.api.ScenarioParametersApiEndpoints.Dtos.ScenarioParametersCombinationWithEngineErrors
+import pl.touk.nussknacker.ui.api.description.ScenarioParametersApiEndpoints
+import pl.touk.nussknacker.ui.api.description.ScenarioParametersApiEndpoints.Dtos.ScenarioParametersCombinationWithEngineErrors
 import pl.touk.nussknacker.ui.process.processingtype.{ProcessingTypeDataProvider, ScenarioParametersService}
 import pl.touk.nussknacker.ui.security.api.{AuthenticationResources, LoggedUser}
