+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/description/AppApiEndpoints.scala
@@ -1,19 +1,18 @@
-package pl.touk.nussknacker.ui.api
+package pl.touk.nussknacker.ui.api.description
 
 import derevo.circe.{decoder, encoder}
 import derevo.derive
 import enumeratum.EnumEntry.Uppercase
 import enumeratum._
 import io.circe.syntax.EncoderOps
-import io.circe.{Codec => CirceCodec, Decoder, Encoder, Json}
+import io.circe.{Decoder, Encoder, Json, Codec => CirceCodec}
 import pl.touk.nussknacker.engine.api.CirceUtil.HCursorExt
 import pl.touk.nussknacker.restmodel.BaseEndpointDefinitions
-import BaseEndpointDefinitions.SecuredEndpoint
+import pl.touk.nussknacker.restmodel.BaseEndpointDefinitions.SecuredEndpoint
 import pl.touk.nussknacker.security.AuthCredentials
 import sttp.model.StatusCode.{InternalServerError, NoContent, Ok}
 import sttp.tapir.EndpointIO.Example
 import sttp.tapir._
-import sttp.tapir.codec.enumeratum._
 import sttp.tapir.derevo.schema
 import sttp.tapir.json.circe.jsonBody
