+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/description/UserApiEndpoints.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.ui.api
+package pl.touk.nussknacker.ui.api.description
 
 import derevo.circe.{decoder, encoder}
 import derevo.derive
